+++ src/constant.h
@@ -138,6 +138,10 @@ constexpr const char *OP_AND = DEFINE_BINARY_OP_NAME("and");
 constexpr const char *OP_OR = DEFINE_BINARY_OP_NAME("or");
 constexpr const char *OP_XOR = DEFINE_BINARY_OP_NAME("xor");
 
+constexpr const char *OP_LSHIFT = DEFINE_BINARY_OP_NAME("<<");
+constexpr const char *OP_RSHIFT = DEFINE_BINARY_OP_NAME(">>");
+constexpr const char *OP_URSHIFT = DEFINE_BINARY_OP_NAME(">>>");
+
 constexpr const char *OP_MATCH = DEFINE_BINARY_OP_NAME("=~");
 constexpr const char *OP_UNMATCH = DEFINE_BINARY_OP_NAME("!~");